Make process-fn-arity a parameter of process-fn-
This makes it possible to reuse process-fn- as I do in the combifn and defcombifn macros as discussed here:
https://groups.google.com/d/msg/prismatic-plumbing/2BEQvdaUIRw/BO8yD_5PQdMJ
For now I copied process-fn- in my code and made that change and was successful in writing these macros while reusing most of the logic in schema.macros. When this PR is merged I'll be able to remove the duplicated process-fn- in my code.