Update README.md
Hello! First of all, thank you guys for this amazing library :)

I'm adding an example of s/conditional to the Readme,
I was struggling with this and I found the feature inside core.cljx but nothing outside to point me out
to the right place.

Let me know if it's inside the right section or if something needs to be changed.